+++ config/webapp.php
@@ -1,0 +1,13 @@
+<?php
+
+/*
+ * For more information, see:
+ * https://github.com/bayfrontmedia/bones-service-webapp/blob/master/docs/setup.md#configuration
+ *
+ * NOTE:
+ * This entire array is added to the Veil data array with key of "webapp"
+ */
+
+return [
+    'version' => '1.0.0', // Web app version
+];